Nuevos archivos independientes para ejercicios

+++ percolar.h
@@ -1,67 +1,67 @@
-extern void llenar(int *red,
-		   int n,
+extern void llenar(long *red,
+		   long n,
 		   double proba,
-		   int *semilla);
+		   long *semilla);
 
-extern int hoshen(int *red,
-		  int n);
+extern long hoshen(long *red,
+		  long n);
 
-extern int actualizar(int *red,
-		      int *clase,
-		      int s,
-		      int frag);
+extern long actualizar(long *red,
+		      long *clase,
+		      long s,
+		      long frag);
 
-extern void etiqueta_falsa(int *red,
-			   int *clase,
-			   int s1,
-			   int s2);
+extern void etiqueta_falsa(long *red,
+			   long *clase,
+			   long s1,
+			   long s2);
 
-extern void corregir_etiqueta(int *red,
-			      int *clase,
-			      int n);
+extern void corregir_etiqueta(long *red,
+			      long *clase,
+			      long n);
 
-extern int percola(int *red,
-		   int n);
+extern long percola(long *red,
+		   long n);
 
-extern int id_percolantes(int *red,
-			  int n,
-			  int *percolantes);
+extern long id_percolantes(long *red,
+			  long n,
+			  long *percolantes);
 
-extern void hist(int *datos,
-		 int *resultado,
-		 int n);
+extern void hist(long *datos,
+		 long *resultado,
+		 long n);
 
-extern void iterar_prob_fija(int n,
-			     int semilla_inicial,
+extern void iterar_prob_fija(long n,
+			     long semilla_inicial,
 			     double proba,
-			     int n_iter,
-			     int *p_total,
+			     long n_iter,
+			     long *p_total,
 			     double *spt_total,
 			     double *spm_total,
 			     double *spmax_total,
 			     double *snp_total,
 			     double *s0_total,
-			     int *np_total,
+			     long *np_total,
 			     double *fppt_total,
 			     double *fppmax_total,
-			     int *ns_total);
+			     long *ns_total);
 
-extern double forzar_percolacion(int n, int semilla,
+extern double forzar_percolacion(long n, long semilla,
 				 double proba_inicial,
-				 int profundidad);
+				 long profundidad);
 
-extern void iterar_buscar_pc(int n,
-			     int semilla_inicial,
-			     int n_iter,
-			     int profundidad,
+extern void iterar_buscar_pc(long n,
+			     long semilla_inicial,
+			     long n_iter,
+			     long profundidad,
 			     double *proba,
 			     double *proba2);
 
-extern void reemplazar(int *red,
-		       int *clase,
-		       int n);
+extern void reemplazar(long *red,
+		       long *clase,
+		       long n);
 
-extern void imprimir(int *red,
-		     int n);
+extern void imprimir(long *red,
+		     long n);
 
-extern double rnd(int *semilla);
+extern double rnd(long *semilla);
